refactored settings service to profile service
