+++ src/services/profile/index.tsx
@@ -1,14 +1,11 @@
 import { createApi } from '@reduxjs/toolkit/query/react'
 import { graphqlRequestBaseQuery } from '@rtk-query/graphql-request-base-query'
-import {
-  GET_USER_PROFILE,
-  POST_USER_PROFILE,
-} from '@/services/settings/queries'
+import { GET_USER_PROFILE, POST_USER_PROFILE } from '@/services/profile/queries'
 import config from '@/config'
-import { ProfileDataType } from '@/types/SettingsType'
+import { ProfileDataType } from '@/types/ProfileType'
 import { GraphQLClient } from 'graphql-request'
 
-export const settingsApiClient = new GraphQLClient(config.graphqlUrl)
+export const profileApiClient = new GraphQLClient(config.graphqlUrl)
 
 type UserProfileData = {
   getProfile: ProfileDataType
@@ -20,17 +17,14 @@ type SetProfileResponse = {
   }
 }
 
-export const settingsApi = createApi({
-  reducerPath: 'settingsApi',
-  baseQuery: graphqlRequestBaseQuery({ client: settingsApiClient }),
+export const profileApi = createApi({
+  reducerPath: 'profileApi',
+  baseQuery: graphqlRequestBaseQuery({ client: profileApiClient }),
   endpoints: builder => ({
     getUserProfile: builder.query<ProfileDataType, string>({
       query: (id: string) => ({
         document: GET_USER_PROFILE,
         variables: { id },
-        headers: {
-          authorization: localStorage.getItem('token'),
-        },
       }),
       transformResponse: (response: UserProfileData) => response.getProfile,
     }),
@@ -54,6 +48,6 @@ export const {
   useGetUserProfileQuery,
   usePostUserProfileMutation,
   util: { getRunningOperationPromises },
-} = settingsApi
+} = profileApi
 
-export const { getUserProfile, postUserProfile } = settingsApi.endpoints
+export const { getUserProfile, postUserProfile } = profileApi.endpoints
