+++ src/store/store.ts
@@ -13,7 +13,7 @@ import { categoriesApi } from '@/services/categories'
 import { activitiesApi } from '@/services/activities'
 import { navSearchApi } from '@/services/search'
 import { tokenStatsApi } from '@/services/token-stats'
-import { settingsApi } from '@/services/settings'
+import { profileApi } from '@/services/profile'
 
 export const store = configureStore({
   reducer: {
@@ -29,7 +29,7 @@ export const store = configureStore({
     [activitiesApi.reducerPath]: activitiesApi.reducer,
     [navSearchApi.reducerPath]: navSearchApi.reducer,
     [tokenStatsApi.reducerPath]: tokenStatsApi.reducer,
-    [settingsApi.reducerPath]: settingsApi.reducer,
+    [profileApi.reducerPath]: profileApi.reducer,
   },
   middleware: gDM =>
     gDM({ serializableCheck: true })
